Refactor by centralizing validations into its own file
