+++ src/runtime/components/fields.tsx
@@ -25,133 +25,18 @@ import {
   type FormPrimitive,
   type SelectValue,
 } from "../../config"
+import {
+  fmeDateTimeToInput,
+  inputToFmeDateTime,
+  fmeDateToInput,
+  inputToFmeDate,
+  fmeTimeToInput,
+  inputToFmeTime,
+  normalizedRgbToHex,
+  hexToNormalizedRgb,
+  normalizeFormValue,
+} from "../../shared/validations"
 import defaultMessages from "./translations/default"
-
-const pad2 = (n: number) => String(n).padStart(2, "0")
-
-const fmeDateTimeToInput = (v: string): string => {
-  // YYYYMMDDHHmmss -> YYYY-MM-DDTHH:mm[:ss]
-  const s = (v || "").replace(/\D/g, "")
-  if (s.length < 12) return ""
-  const y = s.slice(0, 4)
-  const m = s.slice(4, 6)
-  const d = s.slice(6, 8)
-  const hh = s.slice(8, 10)
-  const mm = s.slice(10, 12)
-  const ss = s.length >= 14 ? s.slice(12, 14) : ""
-  return `${y}-${m}-${d}T${hh}:${mm}${ss ? `:${ss}` : ""}`
-}
-
-const inputToFmeDateTime = (v: string): string => {
-  // YYYY-MM-DDTHH:mm[:ss] -> YYYYMMDDHHmmss
-  if (!v) return ""
-  const s = v.trim()
-  const [date, time] = s.split("T")
-  if (!date || !time) return ""
-
-  const [y, m, d] = date.split("-")
-  const [hh, mi, ssRaw] = time.split(":")
-
-  // Year must be 4 digits
-  if (!y || y.length !== 4 || !/^[0-9]{4}$/.test(y)) return ""
-
-  const safePad2 = (part?: string): string | null => {
-    if (!part && part !== "0") return null
-    const n = Number(part)
-    if (!Number.isFinite(n)) return null
-    return pad2(n)
-  }
-
-  const m2 = safePad2(m)
-  const d2 = safePad2(d)
-  const hh2 = safePad2(hh)
-  const mi2 = safePad2(mi)
-  if (!m2 || !d2 || !hh2 || !mi2) return ""
-
-  const ss2 = ssRaw ? safePad2(ssRaw) : "00"
-  if (ss2 === null) return ""
-
-  return `${y}${m2}${d2}${hh2}${mi2}${ss2}`
-}
-
-const fmeDateToInput = (v: string): string => {
-  // YYYYMMDD -> YYYY-MM-DD
-  const s = (v || "").replace(/\D/g, "")
-  if (s.length !== 8) return ""
-  return `${s.slice(0, 4)}-${s.slice(4, 6)}-${s.slice(6, 8)}`
-}
-
-const inputToFmeDate = (v: string): string => (v ? v.replace(/-/g, "") : "")
-// HHmmss or HHmm -> HH:mm[:ss]
-const fmeTimeToInput = (v: string): string => {
-  const s = (v || "").replace(/\D/g, "")
-  if (s.length === 4) return `${s.slice(0, 2)}:${s.slice(2, 4)}`
-  if (s.length >= 6) return `${s.slice(0, 2)}:${s.slice(2, 4)}:${s.slice(4, 6)}`
-  return ""
-}
-
-const inputToFmeTime = (v: string): string => {
-  // HH:mm or HH:mm:ss -> HHmmss
-  if (!v) return ""
-  const parts = v.split(":").map((x) => x || "")
-  const hh = parts[0] || ""
-  const mm = parts[1] || ""
-  const ss = parts[2] || ""
-
-  const nH = Number(hh)
-  const nM = Number(mm)
-  // FME time requires HHmm, ss is optional
-  if (!Number.isFinite(nH) || !Number.isFinite(nM)) return ""
-
-  const nS = Number(ss)
-  const finalSS = Number.isFinite(nS) ? pad2(nS) : "00"
-
-  return `${pad2(nH)}${pad2(nM)}${finalSS}`
-}
-
-const normalizedRgbToHex = (v: string): string | null => {
-  // "r,g,b[,a]" floats (0..1) -> "#RRGGBB"
-  const parts = (v || "").split(",").map((s) => s.trim())
-  if (parts.length < 3) return null
-  const to255 = (f: string) => {
-    const n = Number(f)
-    if (!Number.isFinite(n)) return null
-    const clamped = Math.max(0, Math.min(1, n))
-    return Math.round(clamped * 255)
-  }
-  const r = to255(parts[0])
-  const g = to255(parts[1])
-  const b = to255(parts[2])
-  if (r == null || g == null || b == null) return null
-  const toHex = (n: number) => n.toString(16).padStart(2, "0")
-  return `#${toHex(r)}${toHex(g)}${toHex(b)}`
-}
-
-const hexToNormalizedRgb = (hex: string): string | null => {
-  // "#RRGGBB" -> "r,g,b" floats (0..1)
-  const m = /^#?([0-9a-f]{6})$/i.exec(hex || "")
-  if (!m) return null
-  const n = parseInt(m[1], 16)
-  const r = (n >> 16) & 0xff
-  const g = (n >> 8) & 0xff
-  const b = n & 0xff
-  const f = (x: number) => Number((x / 255).toFixed(6)).toString()
-  return `${f(r)},${f(g)},${f(b)}`
-}
-
-// Utility functions for field handling
-export const normalizeFormValue = (
-  value: FormPrimitive | undefined,
-  isMultiSelect: boolean
-): FormPrimitive | SelectValue => {
-  if (value === undefined || value === null) {
-    return isMultiSelect ? [] : ""
-  }
-  if (isMultiSelect) {
-    return Array.isArray(value) ? value : []
-  }
-  return typeof value === "string" || typeof value === "number" ? value : ""
-}
 
 export const makePlaceholders = (
   translate: (k: string, p?: any) => string,
